backend/requirements.txt
# Web framework
Flask==3.0.0
Flask-CORS>=6.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0

# Google Maps / HTTP
googlemaps==4.10.0
requests==2.31.0

# Report / Excel
openpyxl==3.1.2
reportlab==4.0.7
Pillow>=10.2.0
python-docx==1.1.0
pandas>=2.1.3,<2.3

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2

# Browser automation
playwright>=1.40.0

# Server
gunicorn
